requirements.txt
pyautogen==0.2.18
openai==1.3.9
chromadb==0.4.24
python-dotenv==1.0.0
pandas==2.1.4
schedule==1.2.0
pytz==2024.1
matplotlib==3.7.1
plotly==5.18.0
kaleido==0.2.1
telethon==1.36.0
Pillow==10.0.0
httpx==0.24.1
numpy<2.0
euriai

